+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed reported request timings (they were measured wrong, longer than actual)
 - Fixed `--in-flight-limit` to actually be respected (wasn't implemented correctly)
 - Changed default `in_flight_limit` to `8` (from `16`) as that seemed like a better performing number
+- Implemented all criteria filter types and added `product_variants.yaml`
 
 # v0.3.0
 - Added `associations` entry for schema (used on export only)